Module Slurp_cohttp__Server.Make

Cohttp slurp functor.

Parameters

C : CFG

Signature

val cfg : Ezcmdliner.cfg

The Ezcmdliner configuration.

val port : unit -> int

Returns the port value.

val services : unit -> string list

Returns the services to load.

val server : ?⁠mode:Conduit_lwt_unix.server -> ?⁠callback:(Cohttp_lwt_unix.Server.conn -> Cohttp.Request.t -> Cohttp_lwt.Body.t -> Cohttp_lwt_unix.Server.response_action Lwt.t) -> unit -> unit Lwt.t

server ~mode ~callback () launches the cohttp slurp server using the Conduit mode mode. callback is used to overwrite the default slurp behavior to handle some corner case. If callback raises some exception, the default slurp behavior is used.

val command : unit Lwt.t Ezcmdliner.command

The slurp cohttp command specification.